move all times to f32

+++ taiko-rs-client/src/gameplay/beatmap_structs/timing_point.rs
@@ -3,7 +3,7 @@
 #[derive(Clone, Copy)]
 pub struct TimingPoint {
     /// Start time of the timing section, in milliseconds from the beginning of the beatmap's audio. The end of the timing section is the next timing point's time (or never, if this is the last timing point).
-    pub time: f64,
+    pub time: f32,
     /// This property has two meanings:
     ///     For uninherited timing points, the duration of a beat, in milliseconds.
     ///     For inherited timing points, a negative inverse slider velocity multiplier, as a percentage. For example, -50 would make all sliders in this timing section twice as fast as SliderMultiplier.
@@ -17,7 +17,7 @@ impl TimingPoint {
         // time,beatLength,meter,sampleSet,sampleIndex,volume,uninherited,effects
         // println!("{}", str.clone());
         let mut split = str.split(',');
-        let time = split.next().unwrap().parse::<f64>().unwrap().round();
+        let time = split.next().unwrap().parse::<f32>().unwrap().round();
         let beat_length = split.next().unwrap().parse::<f32>().unwrap();
         let _meter = split.next(); //.unwrap().parse::<u32>().unwrap();
         let _sample_set = split.next(); //.unwrap().parse::<u32>().unwrap();
@@ -41,8 +41,8 @@ impl TimingPoint {
     }
 
 
-    pub fn bpm_multiplier(&self) -> f64 {
+    pub fn bpm_multiplier(&self) -> f32 {
         if self.beat_length > 0.0 {1.0}
-        else {(-self.beat_length as f64).clamp(10.0, 1000.0) / 100.0}
+        else {(-self.beat_length as f32).clamp(10.0, 1000.0) / 100.0}
     }
 }
